requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
google-search-results
newspaper4k
lxml_html_clean
youtube_transcript_api
pdfplumber
streamlit 
python-docx